norminette get_command_path
